Chore:Refactor on tools implementations
